requirements.txt
python-dotenv
requests
pandas
duckdb==1.4.3  # Required for MotherDuck compatibility (as of 2025-05-24)
